Add ability to set the width that the hover text will have
This is set to 60 characters by default (same as Minecraft) but will not apply to hover messages that already contain new line characters to allow manual modification. Also it will try to split between words but might cut very long words if they don't fit. This is done in a way that is a bit more intelligent about trying to keep words together than the Vanilla Minecraft algorithm